# RFC: Speed up the _genEvenOdd calculations