@moble 2018-08-20
Closing this as irrelevant, given Hugo's excellent work #43.